Refactor: 모달 마운트 컴포넌트명 변경 및 재사용성 강화 수정
